refactor(benchmarks): update imports to use @fio/core instead of relative paths
affects: @fio/benchmarks, @fio/chai, @fio/core, @fio/example, @fio/prelude, @fio/website
