+++ packages/benchmarks/Constant.ts
@@ -1,10 +1,9 @@
 /**
  * Created by tushar on 2019-05-09
  */
+import {FIO} from '@fio/core'
 import {Promise} from 'bluebird'
 import * as Fluture from 'fluture'
-
-import {FIO} from '../core/src/main/FIO'
 
 import {RunSuite} from './internals/RunSuite'
