+++ packages/benchmarks/FStream.ts
@@ -5,7 +5,7 @@
 /* tslint:disable */
 import {Suite} from 'benchmark'
 
-import {FIO, UIO} from '../core/src/main/FIO'
+import {FIO, UIO} from '@fio/core'
 
 import {PrintLn} from './internals/PrintLn'
 import {fioRuntime} from './internals/RunSuite'
