+++ packages/benchmarks/package.json
@@ -32,8 +32,10 @@
     "@types/bluebird": "^3.5.28",
     "benchmark": "^2.1.4",
     "bluebird": "^3.7.1",
-    "fluture": "^11.0.3",
-    "@fio/core": "*",
-    "@fio/prelude": "*"
+    "fluture": "^11.0.3"
+  },
+  "dependencies": {
+    "@fio/core": "^0.0.0-development",
+    "@fio/prelude": "^0.0.0"
   }
 }
