+++ packages/example/package.json
@@ -29,5 +29,9 @@
   },
   "devDependencies": {
     "source-map-support": "^0.5.13"
+  },
+  "dependencies": {
+    "@fio/core": "^0.0.0-development",
+    "@fio/prelude": "^0.0.0"
   }
 }
