+++ packages/prelude/package.json
@@ -14,5 +14,8 @@
   },
   "bugs": {
     "url": "https://github.com/tusharmath/fearless-io/issues"
+  },
+  "dependencies": {
+    "@fio/core": "^0.0.0-development"
   }
 }
