+++ packages/website/package.json
@@ -32,5 +32,9 @@
   },
   "bugs": {
     "url": "https://github.com/tusharmath/fearless-io/issues"
+  },
+  "dependencies": {
+    "@fio/core": "^0.0.0-development",
+    "@fio/prelude": "^0.0.0"
   }
 }
